Move exceptions to a separate file
