+++ ruobr_api/exceptions.py
@@ -1,0 +1,13 @@
+class AuthenticationException(Exception):
+    def __init__(self, text):
+        self.text = text
+
+
+class NoChildrenException(Exception):
+    def __init__(self, text):
+        self.text = text
+
+
+class NoSuccessException(Exception):
+    def __init__(self, text):
+        self.text = text
